pyproject.toml
[tool.poetry]
name = "storage-workflows"
version = "0.1.0"
description = ""
authors = ["Ao Chen <ao.chen@doordash.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.6.1"
pytz = "^2022.2.1"
boto3 = "^1.24.54"
PyYAML = "^6.0"
tabulate = "^0.8.10"
requests = "^2.28.1"
slack-sdk = "^3.18.1"
pytimeparse = "^1.1.8"
kubernetes = "^24.2.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
requests-mock = "^1.10.0"
pytest = "^7.2.1"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^7.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
storage-workflows = "storage_workflows.main:app"